Feat. add useGetAllDocuments
